tyr: Modify migrate_from_poi_to_osm to delete_dataset

+++ source/navitiacommon/navitiacommon/models.py
@@ -414,18 +414,21 @@ class Instance(db.Model):
         else:
             raise Exception({'error': 'instance is required'}, 400)
 
-    def delete_dataset(self):
+    def delete_dataset(self, _type):
         result = db.session.query(DataSet, Job) \
             .join(Job) \
-            .filter(DataSet.type == 'poi', Job.instance_id == self.id)\
+            .filter(DataSet.type == _type, Job.instance_id == self.id)\
             .all()
 
+        if not result:
+            return 0
         job_list = {}
         for dataset, job in result:
             # Cascade Delete not working so delete Metric associated manually
             db.session.query(Metric).filter(Metric.dataset_id == dataset.id).delete()
             db.session.delete(dataset)
 
+            # Delete a job without any dataset
             if job.id not in job_list:
                 job_list[job.id] = 1
             else:
@@ -435,6 +438,7 @@ class Instance(db.Model):
                 db.session.delete(job)
 
         db.session.commit()
+        return len(result)
 
     def delete_old_jobs_and_list_datasets(self, time_limit):
         """
